passing exercise list as an argument to ExercisesListScreen.kt
